Remove angular from return component
