+++ config/webpack.js
@@ -159,7 +159,7 @@ module.exports = (env) => {
         inject: false,
         filename: 'return.html',
         template: '!handlebars-loader!src/return.html',
-        chunks: ['vendor', 'authReturn']
+        chunks: ['authReturn']
       }),
 
       new HtmlWebpackPlugin({
@@ -185,7 +185,7 @@ module.exports = (env) => {
       // Extract a stable "vendor" chunk
       new webpack.optimize.CommonsChunkPlugin({
         name: 'vendor',
-        chunks: ['main', 'authReturn'],
+        chunks: ['main'],
         minChunks: function(module) {
           // this assumes your vendor imports exist in the node_modules directory
           return module.context && module.context.indexOf('node_modules') !== -1;
